Migrate apollo client setup to index and refactor query structure
