+++ src/index.js
@@ -3,12 +3,43 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App/App';
 import reportWebVitals from './reportWebVitals';
+import {ApolloClient, 
+  InMemoryCache, 
+  ApolloProvider, 
+  HttpLink, 
+  from,
+  useQuery      
+} from '@apollo/client';
+import {onError} from '@apollo/client/link/error';
 
+const errorLink = onError(({graphqlErrors, networkError}) => {
+  if (graphqlErrors) {
+    graphqlErrors.map(({message, location, path}) => {
+      alert(`Graphql error ${message}`);
+    });
+  }
+}) ;
+
+const link = from([
+  errorLink,
+  new HttpLink({uri:"http://localhost:4500/graphql"}) //This will be the link to our backend server when it is up and running! Ask BE team about if we should keep it local or deployed.
+])
+
+const client = new ApolloClient({
+  cache: new InMemoryCache(),
+  link: link,
+  // fetchOptions: {
+  //   mode: 'no-cors'
+  // }
+})
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <ApolloProvider client={client}>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </ApolloProvider>
+  ,
   document.getElementById('root')
 );
